# Improve Python bundling speed