--- a/.changeset/nine-ghosts-leave.md
+++ b/.changeset/nine-ghosts-leave.md
@@ -1,0 +1,5 @@
+---
+"sst": patch
+---
+
+Improve Python bundling speed